# perf: Improve @babel/types builders 2