--- a/packages/babel-types/src/validators/validate.ts
+++ b/packages/babel-types/src/validators/validate.ts
@@ -3,12 +3,13 @@ import {
   NODE_PARENT_VALIDATIONS,
   type FieldOptions,
 } from "../definitions/index.ts";
+import type { FieldDefinitions } from "../definitions/utils.ts";
 import type * as t from "../index.ts";
 
 export default function validate(
   node: t.Node | undefined | null,
   key: string,
-  val: any,
+  val: unknown,
 ): void {
   if (!node) return;
 
@@ -20,10 +21,31 @@ export default function validate(
   validateChild(node, key, val);
 }
 
+export function validateInternal(
+  defs: FieldDefinitions,
+  node: t.Node | undefined | null,
+  key: string,
+  val: unknown,
+  maybeNode?: 1,
+): void {
+  const field = defs[key];
+
+  if (!field?.validate) return;
+  if (field.optional && val == null) return;
+
+  field.validate(node, key, val);
+
+  if (maybeNode) {
+    const type = (val as t.Node).type;
+    if (type == null) return;
+    NODE_PARENT_VALIDATIONS[type]?.(node, key, val);
+  }
+}
+
 export function validateField(
   node: t.Node | undefined | null,
   key: string,
-  val: any,
+  val: unknown,
   field: FieldOptions | undefined | null,
 ): void {
   if (!field?.validate) return;
@@ -35,10 +57,9 @@ export function validateField(
 export function validateChild(
   node: t.Node | undefined | null,
   key: string,
-  val?: t.Node | undefined | null,
+  val?: unknown,
 ) {
-  if (val == null) return;
-  const validate = NODE_PARENT_VALIDATIONS[val.type];
-  if (!validate) return;
-  validate(node, key, val);
+  const type = (val as t.Node)?.type;
+  if (type == null) return;
+  NODE_PARENT_VALIDATIONS[type]?.(node, key, val);
 }