@nicolo-ribaudo 2024-09-18
> I was surprised that the second commit improved performance by nearly 40%.

Maybe because it makes `validate`'s first argument monomorphic (the new objects all have the same shape)