@liuxingbaoyu 2024-09-18
> It'd be good to have a benchmark that calls multiple different validators.

I added a new benchmark, but something even stranger seems to be happening. 😅
Below are two benchmarks that I consider equivalent, but they give different results. 🤦‍♂️

```
import baseline from "@babel-baseline/types";
import current from "@babel/types";
import { Benchmark } from "../../util.mjs";

const bench = new Benchmark();

function benchCases(implementations) {
  for (const version in implementations) {
    const t = implementations[version];
    const func = t.stringLiteral;
    const func2 = t.memberExpression;
    const func3 = t.assignmentExpression;
    const func4 = t.binaryExpression;

    const id = t.identifier("id");
    const id2 = t.identifier("id2");
    bench.add(`${version} builder`, () => {
      func("bar");
      func2(id, id2, /*computed?*/ false /*, optional? missing*/);
      func3("=", id, id2);
      func4("+", id, id2);
    });
  }
}

benchCases({ baseline, current });

bench.run();
```
```
PS F:\babel\benchmark\babel-types\builders> node .\multiple.mjs
baseline builder: 252_117 ops/sec ±0.19% (0.004ms)
current builder: 2_729_772 ops/sec ±0.48% (0ms)
```
and

```
import { Benchmark, baselineTypes, currentTypes } from "../../util.mjs";

const bench = new Benchmark();

function benchCases(implementations) {
  for (const version in implementations) {
    const t = implementations[version];
    const func = t.stringLiteral;
    const func2 = t.memberExpression;
    const func3 = t.assignmentExpression;
    const func4 = t.binaryExpression;

    const id = t.identifier("id");
    const id2 = t.identifier("id2");
    bench.add(`${version} builder`, () => {
      func("bar");
      func2(id, id2, /*computed?*/ false /*, optional? missing*/);
      func3("=", id, id2);
      func4("+", id, id2);
    });
  }
}

benchCases({ baselineTypes, currentTypes });

bench.run();
```
```
PS F:\babel\benchmark\babel-types\builders> node .\multiple.mjs
baselineTypes builder: 232_489 ops/sec ±0.3% (0.004ms)
currentTypes builder: 4_012_689 ops/sec ±0.31% (0ms)
```