@nicolo-ribaudo 2024-09-23
@liuxingbaoyu The only difference I see is the second one is is importing with `import * as baselineTypes from "@babel-baseline/types";`, while the first one with `import baselineTypes from "@babel-baseline/types";`.

This probably makes a difference because the properties in the first one might be cached, while in the second one they are not?