@liuxingbaoyu 2024-09-23
I remember I tried it and the results were still different.

But today when I tried it again, I found that all benchmarks gave the same results. 🤦‍♂️